refactored Song list, creation and detail code
