+++ PedalCore/Songs/Creation/SongCreation.swift
@@ -1,0 +1,14 @@
+//
+//  SongCreation.swift
+//  PedalCore
+//
+//  Created by Matheus Berger on 30/12/23.
+//
+
+import Foundation
+
+extension Song {
+    struct Creation {
+        // Namespace umbrella struct for SongCreation related code
+    }
+}
