+++ PedalCore/Songs/Details/SongDetails.swift
@@ -1,0 +1,14 @@
+//
+//  SongDetails.swift
+//  PedalCore
+//
+//  Created by Matheus Berger on 30/12/23.
+//
+
+import Foundation
+
+extension Song {
+    struct Details {
+        // Namespace umbrella struct for SongDetails related code
+    }
+}
