+++ PedalCore/Songs/List/SongList.swift
@@ -1,0 +1,14 @@
+//
+//  SongList.swift
+//  PedalCore
+//
+//  Created by Matheus Berger on 30/12/23.
+//
+
+import Foundation
+
+extension Song {
+    struct List {
+        // Namespace umbrella struct for SongList related code
+    }
+}
