+++ PedalCore/Songs/Models/Song+Populate.swift
@@ -8,16 +8,16 @@
 import Foundation
 
 extension Song {
-    static func getSample() -> [Song] {
+    static func getSample() -> [Song.Model] {
         return [
-            Song(name: "Paranoid Android",
+            Model(name: "Paranoid Android",
                  artist: "Radiohead",
                  pedals: [Pedal.pedalSample()[0],
                           Pedal.pedalSample()[1],
                           Pedal.pedalSample()[2]
                          ]
                 ),
-            Song(name: "Teddy Picker",
+            Model(name: "Teddy Picker",
                  artist: "ArcticMonkeys",
                  pedals: [Pedal.pedalSample()[2],
                           Pedal.pedalSample()[3]
