+++ PedalCore/Songs/Song.swift
@@ -1,0 +1,12 @@
+//
+//  Song.swift
+//  PedalCore
+//
+//  Created by Lucas Migge on 15/11/23.
+//
+
+import Foundation
+
+struct Song {
+    // Namespace umbrella struct for Song related code
+}
